Widen the FCI histogram bins EUMETCastVideo counts into
VideoMaker's mtg_histogram was quint16, and one chunk of one band can put
far more than the 65535 that counts to into a single bin. The night side
is the case that bites an animation : with the disc dark, a whole FDHSI
solar channel collapses into three or four bins. At 2026-09-02 00:00 the
wrap threw away 96 % of the counts - 738814 of 28395006 survived - and
the total the cumulative LUT is scaled by went with them.

What it comes to in a finished video is the day side drifting in
brightness across the sequence. The error is proportional to how much of
the disc is dark, so it is worst at the terminator and gone by midday.
Taking whole discs of 2026-09-02 through the per chunk histogram and the
LUT by hand, vis_06 at 05:00 gives 40.6 million counts against 92.5 and
maxRadianceIndex 1237 against 1057 ; at 07:00, 68.9 against 92.5 and 1229
against 1162 ; at 08:00, 80.4 against 92.5 and 1366 against 1326 ; at
12:00 nothing overflows and the two agree exactly.

A VideoMaker is 2.54 MB now rather than 1.29, of which mtg_histogram is
2.5 MB. It is already on the heap ; the comment in main.cpp saying why
carried the old numbers, and one object over the 2 MB a MinGW executable
reserves for its stack is a plainer reason than the old one.

Verified by rendering frame 001 of bin/EUMETCastVideo.json with a build
of each. totpixels[0] comes back 738814 with quint16 and 28395006 with
quint32, matching active_pixels exactly. Frames 001 and 073 are byte
identical between the two builds, which is the no regression check : 073
is full daylight and never overflowed, and 001 is dark enough that every
pixel of that projection comes from the night channel, whose histogram
did not overflow either.

+++ video/videomaker.h
@@ -103,7 +103,14 @@ private:
     quint16 mtg_stat_max[4][40] = {};
     long mtg_active_pixels[4][40] = {};
 
-    quint16 mtg_histogram[4][40][4096] = {};
+    // Millions of pixels of one chunk can land in a single bin, so the bins have
+    // to be wider than the 65535 a quint16 counts to. The night side is the case
+    // that bites an animation : with the disc dark, a whole FDHSI solar channel
+    // collapses into three or four bins, and the wrap threw away 96 % of the
+    // counts. That skews the cumulative LUT and with it the 95 % stretch the
+    // video is made with, worst at the terminator and fading as the day fills
+    // the disc - so the day side drifts in brightness across the sequence.
+    quint32 mtg_histogram[4][40][4096] = {};
 
     quint16 *ptrMTG[4][40] = {};
